+++ Runtime/Tweens/Components/TweenStopCommand.cs
@@ -1,0 +1,6 @@
+﻿using Unity.Entities;
+
+namespace Timespawn.EntityTween.Tweens
+{
+    internal struct TweenStopCommand : IComponentData {}
+}